// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  STUDENT: 'STUDENT',
  LECTURER: 'LECTURER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AccountStatus = {
  ACTIVE: 'ACTIVE',
  INVITED: 'INVITED',
  SUSPENDED: 'SUSPENDED',
  DELETED: 'DELETED'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const CourseStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type CourseStatus = (typeof CourseStatus)[keyof typeof CourseStatus]


export const EnrollmentStatus = {
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  DROPPED: 'DROPPED'
} as const

export type EnrollmentStatus = (typeof EnrollmentStatus)[keyof typeof EnrollmentStatus]


export const SubmissionStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  GRADED: 'GRADED',
  RETURNED: 'RETURNED'
} as const

export type SubmissionStatus = (typeof SubmissionStatus)[keyof typeof SubmissionStatus]


export const FileType = {
  PROFILE_IMAGE: 'PROFILE_IMAGE',
  COURSE_THUMBNAIL: 'COURSE_THUMBNAIL',
  LESSON_VIDEO: 'LESSON_VIDEO',
  LESSON_RESOURCE: 'LESSON_RESOURCE',
  SUBMISSION_ATTACHMENT: 'SUBMISSION_ATTACHMENT',
  DOCUMENT: 'DOCUMENT',
  OTHER: 'OTHER'
} as const

export type FileType = (typeof FileType)[keyof typeof FileType]


export const NotificationType = {
  COURSE_ENROLLMENT: 'COURSE_ENROLLMENT',
  LESSON_PUBLISHED: 'LESSON_PUBLISHED',
  ASSIGNMENT_GRADED: 'ASSIGNMENT_GRADED',
  ANNOUNCEMENT: 'ANNOUNCEMENT',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const AuditAction = {
  USER_CREATED: 'USER_CREATED',
  USER_INVITED: 'USER_INVITED',
  USER_UPDATED: 'USER_UPDATED',
  USER_DELETED: 'USER_DELETED',
  USER_LOGIN: 'USER_LOGIN',
  USER_LOGOUT: 'USER_LOGOUT',
  COURSE_CREATED: 'COURSE_CREATED',
  COURSE_UPDATED: 'COURSE_UPDATED',
  COURSE_DELETED: 'COURSE_DELETED',
  COURSE_PUBLISHED: 'COURSE_PUBLISHED',
  ENROLLMENT_CREATED: 'ENROLLMENT_CREATED',
  ENROLLMENT_UPDATED: 'ENROLLMENT_UPDATED',
  ENROLLMENT_DELETED: 'ENROLLMENT_DELETED',
  LESSON_CREATED: 'LESSON_CREATED',
  LESSON_UPDATED: 'LESSON_UPDATED',
  LESSON_DELETED: 'LESSON_DELETED',
  LESSON_PUBLISHED: 'LESSON_PUBLISHED',
  SUBMISSION_CREATED: 'SUBMISSION_CREATED',
  SUBMISSION_UPDATED: 'SUBMISSION_UPDATED',
  SUBMISSION_GRADED: 'SUBMISSION_GRADED',
  FILE_UPLOADED: 'FILE_UPLOADED',
  FILE_DOWNLOADED: 'FILE_DOWNLOADED',
  FILE_DELETED: 'FILE_DELETED',
  SYSTEM_ERROR: 'SYSTEM_ERROR',
  SYSTEM_WARNING: 'SYSTEM_WARNING'
} as const

export type AuditAction = (typeof AuditAction)[keyof typeof AuditAction]